client/requirements.txt
eventlet==0.25.1
python-socketio==4.4.0
requests==2.22.0